requirements.txt
Flask
numpy
opencv-python-headless
deepface
tensorflow
keras
mysql-connector-python